Added 3 rooms, 2 items, saving system, timer, damage areas, item pickups
Room 0 : Save point (hourglass), door with a keypad. If the player has item "doorcode", going through the keypad will open the door, which will be temporarily saved (new system!) has opened so that it stays opened when you get back to the room.
Room 1 : A corridor with some damaging water. If you go into the water, the player play a hurt animation and get stunned for 0.7s (New State! "Locked") and then get 1.5s of invulnerability time. There is a DamageArea system for the water, reusable.
Room 2 : Simple room with the "doorcode" item (new system, ItemPickupArea). Once picked, immediately gets saved to player.abilities.
Room 3 : Simple room with the "sandcanister_one" item. New variant for ItemPickupArea : this time it gets temporarily savec to player.tempabilities. If the player can get back to the hourglass (start) with it, then it get actually saved (to abilities).

Timer : Added a 00:20 timer at the top of the screen (upgradable). If the player bring back sand canister, this get upgraded (currently set to 0:20 -> 1:00 -> 2:00 -> 4:00 as placeholder). Need to impelment that when it reaches 00:00, the player either die and restart or is rewinded to the center. Maybe the player explode, then play it backward and play a second of rewinding the player toward the center, then fade to black and restart.